+++ packages/react-components/src/InputAddress/index.tsx
@@ -314,15 +314,25 @@ class InputAddress extends React.PureComponent<Props, State> {
       const accountId = transformToAccountId(query);
 
       if (accountId) {
-        const item = keyring.saveRecent(
-          accountId.toString()
-        ).option;
+        const account = keyring.getAccount(accountId);
 
-        matches.push({
-          key: item.key,
-          name: item.name,
-          value: item.value || undefined
-        });
+        if (account) {
+          matches.push({
+            key: account.address,
+            name: account.meta.name,
+            value: account.address
+          });
+        } else {
+          const item = keyring.saveRecent(
+            accountId.toString()
+          ).option;
+
+          matches.push({
+            key: item.key,
+            name: item.name,
+            value: item.value || undefined
+          });
+        }
       }
     }